Fixed #28 作成者や更新日を横断的に設定できるように修正、バージョンが付与されるように修正、コメントを修正

+++ hello-wts/src/main/java/jp/ne/hyukke/wts/hello/domain/model/Sample.java
@@ -15,7 +15,7 @@ import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
 
 /**
- * サンプルのエンティティクラス.
+ * サンプルのドメインモデルクラス.
  *
  * @author hyukke
  */
@@ -38,7 +38,7 @@ public class Sample implements Serializable {
     @DateTimeFormat(pattern = "yyyy/MM/dd HH:mm:ss")
     private LocalDateTime lastUpdatedDate;
     /** バージョン */
-    private long version;
+    private int version;
 
     public Sample() {
         this.id = Integer.valueOf(0);
